% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.optcheck.R
\name{diff.optcheck}
\alias{diff.optcheck}
\alias{diff.optproj}
\alias{diff.optrefit}
\alias{diff.summary.optproj}
\alias{diff.summary.optrefit}
\title{Elementwise difference between potential and optimal solutions.}
\usage{
\method{diff}{optcheck}(x)
}
\arguments{
\item{x}{An object of class \code{optproj}, \code{optrefit}, \code{summary.optproj}, or \code{summary.optrefit}.}
}
\value{
The minimum of the absolute and relative difference between each element of the potential solution (\code{xsol}) and the optimial solution (\code{xopt}) obtained from either \code{\link{optim_proj}} or \code{\link{optim_refit}}.
}
\description{
Elementwise difference between potential and optimal solutions.
}
