% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_check.R
\name{optim_check}
\alias{optim_check}
\title{Numerical check for MLE.}
\usage{
optim_check(loglik, theta.mle, itheta, theta.names, theta.rng, refit = TRUE,
  layout, plot = TRUE)
}
\arguments{
\item{loglik}{loglikelihood function.  Takes a single vector argument.}

\item{theta.mle}{The potential MLE.}

\item{itheta}{indices of one dimensional functions to evaluate and plot.  Defaults to all parameters.}

\item{theta.names}{Optional vector of parameter names for plotting.}

\item{theta.rng}{Optional two-row matrix giving the plotting limits for each parameter.  Defaults to theta.mle +/- .5 * abs(theta.mle)}

\item{refit}{If \code{TRUE}, recalculates the range so that drop is more or less the same on either side of \code{theta.mle}.}

\item{layout}{Optional vector giving the number of rows and columns in the plot.  For \code{p} parameters, defaults to \code{c(nr, nc)}, where \code{nr = floor(p)} and \code{nc = ceiling(p/nr)}.}

\item{plot}{Logical, whether or not to plot the conditional likelihoods.  See Value for details.}
}
\value{
A logical vector of length \code{p}, indicating whether or not the MLE was the maximum of the plotted points along each coordinate.  This is for numerically checking what the plots otherwise do visually.
}
\description{
Given a log-likelihood function and a potential MLE, checks whether each one-dimensional version of the log-likelihood which varies one parameter at a time with all others at the MLE is indeed maximized at the MLE value.
}
