% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.optimCheck.R
\name{plot.optimCheck}
\alias{plot.optimCheck}
\title{Projection plots for optimization routines.}
\usage{
\method{plot}{optimCheck}(x, theta.names, itheta, layout)
}
\arguments{
\item{x}{An \code{optimCheck} object, i.e., output from function \code{\link{optim_check}}.}

\item{theta.names}{Optional vector of parameter names for plotting.}

\item{itheta}{indices of one dimensional functions to evaluate and plot.  Defaults to all parameters.}

\item{layout}{Optional vector giving the number of rows and columns in the plot.  For \code{ntheta} parameters, defaults to \code{c(nr, nc)}, where \code{nr = floor(ntheta)} and \code{nc = ceiling(ntheta/nr)}.}
}
\value{
A grid of projection plots, with vertical lines at the potential solution.
}
\description{
Projection plots for optimization routines.
}
