% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.opt_proj.R
\name{summary.opt_proj}
\alias{summary.opt_proj}
\alias{print.opt_proj}
\alias{print.summary.opt_proj}
\alias{print.summary.opt_proj}
\alias{print.opt_proj}
\title{\code{summary} method for projection plots.}
\usage{
\method{summary}{opt_proj}(object, xnames)

\method{print}{summary.opt_proj}(x, digits = max(3L, getOption("digits") -
  3L))

\method{print}{opt_proj}(x, digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{object}{An \code{opt_proj} object, i.e., output from the function \code{\link{optim_proj}}.}

\item{xnames}{Optional vector of names for the elements of the potential solution.}
}
\value{
A list with elements:
\describe{
  \item{\code{xsol}}{The potential solution vector.}
  \item{\code{ysol}}{The value of the objective function at \code{x}.}
  \item{\code{opt}}{A two-row matrix of optimal values of the projection plots.  The first row corresponds to the maxizing/minimizing value of \code{x_i}, and the second is the value of the maximum/minimum itself.}
  \item{\code{diff}}{A two-row matrix of differences between the potential solution and the optimal value.  The first row corresponds to differences in each element of theta, and the second to differences in the objective function.}
  \item{\code{rel}}{A two-row matrix of relative differences between the potential solution and the optimal value (the denominator being the potential solution).  The first row corresponds to relative differences in each element of theta, and the second to relative differences in the objective function.}
}
}
\description{
\code{summary} method for projection plots.
}
\details{
The \code{print} methods for the summary and \code{opt_proj} objects themselves both return a three-column matrix, consisting of the potential solution, the optimal solution in each projection plot, and the relative difference between the two.
}
