% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.opt_refit.R
\name{summary.opt_refit}
\alias{summary.opt_refit}
\alias{print.opt_reft}
\alias{print.summary.opt_refit}
\alias{print.summary.opt_refit}
\alias{print.opt_refit}
\title{\code{summary} method for \code{opt_refit} objects.}
\usage{
\method{summary}{opt_refit}(object, xnames)

\method{print}{summary.opt_refit}(x, digits = max(3L, getOption("digits") -
  3L))

\method{print}{opt_refit}(x, digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{object}{An \code{opt_refit} object, i.e., output from the function \code{\link{optim_refit}}.}

\item{xnames}{Optional vector of names for the elements of the potential solution.}
}
\value{
A list with elements:
\describe{
  \item{\code{xsol}}{The potential solution vector.}
  \item{\code{ysol}}{The value of the objective function at \code{xsol}.}
  \item{\code{xopt}}{A vector containing the argmax/argmin in each projection plot.}
  \item{\code{yopt}}{A vector containing the max/min in each projection plot.}
  \item{\code{xdiff}}{A two-row matrix containing the differences between \code{xsol} and \code{xopt}.  The first row is the absolute difference \code{D = xopt - xsol}, the second is the relative difference \code{R = D/|xsol|}.}
  \item{\code{ydiff}}{Same thing, but between \code{ysol} and \code{yopt}.}
}
}
\description{
\code{summary} method for \code{opt_refit} objects.
}
