% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.optrefit.R
\name{summary.optrefit}
\alias{summary.optrefit}
\alias{print.optrefit}
\alias{print.summary.optrefit}
\title{\code{summary} method for \code{optrefit} objects.}
\usage{
\method{summary}{optrefit}(object, xnames)
}
\arguments{
\item{object}{An \code{optrefit} object, i.e., output from the function \code{\link{optim_refit}}.}

\item{xnames}{Optional vector of names for the elements of the potential solution.}
}
\value{
A list with elements:
\describe{
  \item{\code{xsol}}{The potential solution vector.}
  \item{\code{ysol}}{The value of the objective function at \code{xsol}.}
  \item{\code{xopt}}{A vector containing the argmax/argmin in each projection plot.}
  \item{\code{yopt}}{A vector containing the max/min in each projection plot.}
  \item{\code{xdiff}}{A two-column matrix containing the differences between \code{xsol} and \code{xopt}.  The first column is the absolute difference \code{D = xopt - xsol}, the second is the relative difference \code{R = D/|xsol|}.}
  \item{\code{ydiff}}{A length-two vector containing the absolute and relative difference between \code{ysol} and \code{yopt}.}
}
}
\description{
\code{summary} method for \code{optrefit} objects.
}
