% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_check.R
\name{optim_check}
\alias{optim_check}
\title{Check solution of an optimization routine.}
\usage{
optim_check(theta, fun, theta.rng = 0.1, refit = TRUE)
}
\arguments{
\item{theta}{Potential solution vector of length \code{ntheta}.}

\item{fun}{Objective function to be maximized (or minimized), with first argument the length-\code{ntheta} parameter vector over which optimization is to take place.  Should return a scalar result.}

\item{theta.rng}{Optional specification of the range of each projection plot.  Can be: (i) a \code{2 x ntheta} matrix giving the endpoints of the range, (ii) a scalar or vector of length \code{ntheta}, such that the range in each plot is \code{theta +/- theta.range * abs(theta)}.}

\item{refit}{If \code{TRUE}, narrows the range of each plot so that the y-value in each plot is more or less the same at either endpoint.}
}
\value{
An object of class \code{optimCheck} consisting of the elements:
\describe{
  \item{\code{theta}}{The potential solution.}
  \item{\code{value}}{The value of \code{fun(theta)}.}
  \item{\code{x}}{An \code{npts x ntheta} matrix where each column is the \code{x}-axis of the projection plot along the given component of \code{theta}.}
  \item{\code{y}}{An \code{npts x ntheta} matrix where each column is the \code{y}-axis of the corresponding projection plot.}
}
}
\description{
Given the objective function of an optimization problem and a potential solution, calculates "projection plots" along each coordinate of the solution vector, with all other coordinates being fixed at the input values.
}
