% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_check.R
\name{optim_check}
\alias{optim_check}
\title{Check solution of an optimization routine.}
\usage{
optim_check(theta, fun, itheta, theta.names, theta.rng, refit = TRUE, layout,
  plot = TRUE)
}
\arguments{
\item{theta}{Value of potential solution vector.}

\item{fun}{Objective function to be maximized (or minimized), with first argument the parameter vector over which optimization is to take place.  Should return a scalar result.}

\item{itheta}{indices of one dimensional functions to evaluate and plot.  Defaults to all parameters.}

\item{theta.names}{Optional vector of parameter names for plotting.}

\item{theta.rng}{Optional two-row matrix giving the plotting limits for each parameter.  Defaults to theta.mle +/- .5 * abs(theta.mle)}

\item{refit}{If \code{TRUE}, recalculates the range so that drop is more or less the same on either side of \code{theta.mle}.}

\item{layout}{Optional vector giving the number of rows and columns in the plot.  For \code{p} parameters, defaults to \code{c(nr, nc)}, where \code{nr = floor(p)} and \code{nc = ceiling(p/nr)}.}

\item{plot}{Logical, whether or not to plot the conditional likelihoods.  See Value for details.}
}
\value{
An object of class \code{optimCheck} consisting of the elements:
\describe{
  \item{\code{theta}}{The potential solution.}
  \item{\code{value}}{The value of \code{fun(theta)}.}
  \item{\code{x}}{An \code{npts x ntheta} matrix where each column is the \code{x}-axis of the projection plot along the given component of \code{theta}.}
  \item{\code{y}}{An \code{npts x ntheta} matrix where each column is the \code{y}-axis of the corresponding projection plot.}
}
The object is returned invisibly if \code{plot = TRUE}.
}
\description{
Given a log-likelihood function and a potential MLE, checks whether each one-dimensional version of the log-likelihood which varies one parameter at a time with all others at the MLE is indeed maximized at the MLE value.
}
