% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimCheck-package.R
\docType{package}
\name{optimCheck}
\alias{optimCheck}
\alias{optimCheck-package}
\title{Graphical and numerical checks for mode-finding routines.}
\description{
Graphical and numerical checks for mode-finding routines.
}
