% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.opt_proj.R
\name{summary.opt_proj}
\alias{summary.opt_proj}
\alias{print.opt_proj}
\alias{print.summary.opt_proj}
\alias{print.summary.opt_proj}
\alias{print.opt_proj}
\title{\code{summary} method for projection plots.}
\usage{
\method{summary}{opt_proj}(object, xnames)

\method{print}{summary.opt_proj}(x, digits = max(3L, getOption("digits") -
  3L))

\method{print}{opt_proj}(x, digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{object}{An \code{opt_proj} object, i.e., output from the function \code{\link{optim_proj}}.}

\item{xnames}{Optional vector of names for the elements of the potential solution.}
}
\value{
A list with elements:
\describe{
  \item{\code{xsol}}{The potential solution vector.}
  \item{\code{ysol}}{The value of the objective function at \code{xsol}.}
  \item{\code{xopt}}{A vector containing the argmax/argmin in each projection plot.}
  \item{\code{yopt}}{A vector containing the max/min in each projection plot.}
  \item{\code{xdiff}}{A two-column matrix containing the differences between \code{xsol} and \code{xopt}.  The first column is the absolute difference \code{D = xopt - xsol}, the second is the relative difference \code{R = D/|xsol|}.}
  \item{\code{ydiff}}{Same thing, but between \code{ysol} and \code{yopt}.}
}
}
\description{
\code{summary} method for projection plots.
}
\details{
The \code{print} methods for the summary and \code{opt_proj} objects themselves both return a three-column matrix, consisting of the potential solution (\code{xsol}), the optimal solution in each projection plot (\code{xopt}), and the relative difference between the two (\code{R = (xopt - xsol)/|xsol|}).
}
